# Handwritten pickle-unpickle operations