--- a/cache/carbonlink_test.go
+++ b/cache/carbonlink_test.go
@@ -13,7 +13,7 @@ import (
 )
 
 const sampleCacheQuery = "\x00\x00\x00Y\x80\x02}q\x01(U\x06metricq\x02U,carbon.agents.carbon_agent_server.cache.sizeq\x03U\x04typeq\x04U\x0bcache-queryq\x05u."
-const sampleCacheQuery2 = "\x00\x00\x00Y\x80\x02}q\x01(U\x06metricq\x02U,carbon.agents.carbon_agent_server.param.sizeq\x03U\x04typeq\x04U\x0bcache-queryq\x05u."
+const sampleCacheQuery2 = "\x00\x00\x00Y\x80\x02}q\x01(U\x04typeq\x04U\x0bcache-queryq\x05U\x06metricq\x02U,carbon.agents.carbon_agent_server.param.sizeq\x03u."
 
 func TestCarbonlink(t *testing.T) {
 	assert := assert.New(t)
@@ -278,8 +278,14 @@ func TestParseCarbonlinkRequest(t *testing.T) {
 				Type:   "cache-query",
 				Metric: "carbon.agents.carbon_agent_server.cache.size"},
 		},
+		{name: "Good query2",
+			data: []byte(sampleCacheQuery2)[4:],
+			want: &CarbonlinkRequest{
+				Type:   "cache-query",
+				Metric: "carbon.agents.carbon_agent_server.param.size"},
+		},
 		{name: "Invalid query type",
-			data: []byte("\x80\x02}q\x00(U\x06Metricq\x01U\x03barq\x02U\x04Typeq\x03U\x03fooq\x04u."),
+			data: []byte("\x80\x02}q\x00(U\x06metricq\x01U\x03barq\x02U\x04typeq\x03U\x03fooq\x04u."),
 			want: &CarbonlinkRequest{
 				Type:   "foo",
 				Metric: "bar",